+++ lib/api_watch.ml
@@ -4,3 +4,4 @@ module Text_diff = Text_diff
 module Library = Library
 module Normalize = Normalize
 module Stddiff = Stddiff
+module Ocaml_types = Ocaml_types